+++ java-service-examples/src/main/java/nbbrd/service/examples/HashAlgorithm.java
@@ -14,7 +14,7 @@ import java.util.stream.Stream;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-@ServiceDefinition(quantifier = Quantifier.MULTIPLE, batch = true)
+@ServiceDefinition(quantifier = Quantifier.MULTIPLE, batchType = HashAlgorithm.Batch.class)
 public interface HashAlgorithm {
 
     // 💡 Enforce service naming
@@ -33,8 +33,12 @@ public interface HashAlgorithm {
                 .ifPresent(System.out::println);
     }
 
+    interface Batch {
+        Stream<HashAlgorithm> getProviders();
+    }
+
     @ServiceProvider
-    class MessageDigestBridge implements HashAlgorithmBatch {
+    class MessageDigestBridge implements Batch {
 
         @Override
         public Stream<HashAlgorithm> getProviders() {