+++ java-service-processor/src/test/resources/definition/TestBatchInvalidType.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+
+import java.util.stream.Stream;
+
+@ServiceDefinition(batchType = TestBatchInvalidType.SomeBatch.class)
+interface TestBatchInvalidType {
+
+    class SomeBatch {
+
+        Stream<TestBatchInvalidType> getProviders();
+    }
+}