+++ java-service-processor/src/test/resources/definition/TestBatchValidType.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+
+import java.util.stream.Stream;
+
+@ServiceDefinition(batchType = TestBatchValidType.SomeBatch.class)
+interface TestBatchValidType {
+
+    interface SomeBatch {
+
+        Stream<TestBatchValidType> getProviders();
+    }
+}