+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add mustache templating in loader and batch names [#320](https://github.com/nbbrd/java-service-util/issues/320)
+- Add explicit declaration of batch class [#216](https://github.com/nbbrd/java-service-util/issues/216)
 
 ### Changed